mechanic fore person auth, updating job price by hour rate

+++ garits/src/main/java/com/garits/user/Role.java
@@ -14,7 +14,7 @@ public class Role {
     @Column(name = "role_name")
     private String roleName;
     @Column(name="hourly_rate")
-    private int hourlyRate;
+    private Double hourlyRate;
     // CONSTRUCTORS
     public Role() {
     }
@@ -27,7 +27,7 @@ public class Role {
         this.idRole = idRole;
     }
 
-    public Role(Integer idRole, String roleName, int hourlyRate) {
+    public Role(Integer idRole, String roleName, Double hourlyRate) {
         this.idRole = idRole;
         this.roleName = roleName;
         this.hourlyRate = hourlyRate;
@@ -51,13 +51,13 @@ public class Role {
     public void setRoleName(String roleName) {
         this.roleName = roleName;
     }
-    
 
-    public int getHourlyRate() {
+
+    public Double getHourlyRate() {
         return hourlyRate;
     }
 
-    public void setHourlyRate(int hourlyRate) {
+    public void setHourlyRate(Double hourlyRate) {
         this.hourlyRate = hourlyRate;
     }
 }
